backend/pyproject.toml
[tool.poetry]
name = "k-matcher"
version = "0.1.0"
description = ""
authors = ["Sleptsov Denis <hurrdurrrderp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.2.1"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
sqlmodel = "^0.0.24"
pyyaml = "^6.0.2"
msgspec = "^0.19.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.9.0"
bandit = "^1.7.8"
pytest = "^8.1.1"
pre-commit = "^3.7.0"
flake8-pyproject = "^1.2.3"
types-pyyaml = "^6.0.12.20240311"
httpx = "^0.28.1"
pytest-asyncio = "^1.1.0"

[tool.bandit]
exclude_dirs = [
    "__pycache__",
    ".mypy_cache",
    ".venv"
]

[tool.black]
line-length = 100
skip-string-normalization = true

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".mypy_cache",
    ".venv"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
